+++ stroom-dashboard/stroom-storedquery-impl-db/src/main/resources/stroom/storedquery/impl/db/migration/V07_00_00_001__query.sql
@@ -32,7 +32,7 @@ CREATE TABLE IF NOT EXISTS query (
     component_id          varchar(255) NOT NULL,
     name                  varchar(255) NOT NULL,
     data                  longtext,
-    favourite             bit(1) NOT NULL,
+    favourite             tinyint(1) NOT NULL DEFAULT '0',
     PRIMARY KEY           (id)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;